+++ include/blickfeld/scanner.h
@@ -18,6 +18,7 @@
 
 #include "blickfeld/scanner-config.h"
 #include "blickfeld/data.h"
+#include "blickfeld/point_cloud_stream.h"
 
 #ifndef BSL_STANDALONE
 #include "blickfeld/connection.pb.h"
@@ -34,26 +35,13 @@ class context;
     #endif
 }
 
-namespace google {
-namespace protobuf {
-namespace io {
-class IstreamInputStream;
-class GzipInputStream;
-class CodedInputStream;
-}
-}
-}
-
 namespace blickfeld {
 
 namespace protocol {
 class Response;
+class Response_Hello;
 namespace data {
 class Frame;
-}
-namespace file {
-class PointCloud_Metadata;
-class PointCloud_Data;
 }
 namespace config {
 class ScanPattern_Filter;
@@ -69,113 +57,21 @@ class point_cloud_record;
 /**
  * Blickfeld Scanner class for point cloud streams, status, and configuration requests.
  */
-class scanner : public logged_object {
+class BF_DLLEXPORT scanner : public logged_object {
 public:
 
 #ifndef BSL_STANDALONE
-	const protocol::data::Frame REF_FRAME_XYZ; // Reference frame: XYZ coordinates
-	const protocol::data::Frame REF_FRAME_XYZ_I; // Reference frame: XYZ coordinates, intensity
-	const protocol::data::Frame REF_FRAME_XYZ_I_ID; // Reference frame: XYZ coordinates, intensity, frame id, scanline id, point id, return id
-	const protocol::data::Frame REF_FRAME_XYZ_I_ID_TS; // Reference frame: XYZ coordinates, intensity, frame id, scanline id, point id, return id, timestamps
-	const protocol::data::Frame REF_FRAME_DEPTH_MAP; // Reference frame: ambient_light_level, intensity, range, frame id, scanline id, point id
+	static const protocol::data::Frame REF_FRAME_XYZ; // Reference frame: XYZ coordinates
+	static const protocol::data::Frame REF_FRAME_XYZ_I; // Reference frame: XYZ coordinates, intensity
+	static const protocol::data::Frame REF_FRAME_XYZ_I_ID; // Reference frame: XYZ coordinates, intensity, frame id, scanline id, point id, return id
+	static const protocol::data::Frame REF_FRAME_XYZ_I_ID_TS; // Reference frame: XYZ coordinates, intensity, frame id, scanline id, point id, return id, timestamps
+	static const protocol::data::Frame REF_FRAME_DEPTH_MAP; // Reference frame: ambient_light_level, intensity, range, frame id, scanline id, point id
+	static const protocol::data::Frame REF_FRAME_FULL; // Reference frame: All attributes (default)
+	static const protocol::data::Frame REF_FRAME_PACKED; // Reference frame in packed format: XYZ coordinates, direction, range, intensity, ambient_light_level, frame id, scanline id, point id, return id, timestamps
 #endif
 
-	/**
-	 * Point Cloud Stream class for requesting and recording Blickfeld Point Clouds.
-	 *
-	 * Use scanner::get_point_cloud_stream or scanner::get_simple_point_cloud_stream for live streams.
-	 * Use the static methods scanner::file_point_cloud_stream or scanner::simple_file_point_cloud_stream for reading recorded point clouds.
-	 */
 	template<class frame_t>
-	class point_cloud_stream {
-		std::shared_ptr<connection> conn;
-		protocol::Response* resp = nullptr;
-		protocol::file::PointCloud_Metadata* metadata = nullptr;
-		frame_t* frame = nullptr;
-
-#ifdef BSL_RECORDING
-		protocol::file::PointCloud_Data* stream_data = nullptr;
-		bool stream_buffered = false;
-		void read_stream();
-
-		std::istream* istream = nullptr;
-		google::protobuf::io::IstreamInputStream* pb_istream = nullptr;
-		google::protobuf::io::GzipInputStream* pb_izstream = nullptr;
-		google::protobuf::io::CodedInputStream* pb_icstream = nullptr;
-		std::ostream* ostream = nullptr;
-		point_cloud_record* record = nullptr;
-#endif
-
-public:
-		/**
-		 * Internal use. Use scanner::get_point_cloud_stream or scanner::get_simple_point_cloud_stream.
-		 *
-		 * @param conn Scanner connection which should be used for stream.
-		 * @param filter Filter points and returns by point attributes during the post-processing on the device.
-		 * @param reference_frame Frame representing the desired data. To request a field, set it to any value (also in submessages). For a repeated field, add at least one element.
-		 * @param extend_subscription Template subscription. Enabled extenstions from this parameter are used to open the stream.
-		 */
-		point_cloud_stream(std::shared_ptr<connection> conn, const protocol::config::ScanPattern_Filter* filter = nullptr, const protocol::data::Frame* reference_frame = nullptr, const protocol::stream::Subscribe_PointCloud* extend_subscription = nullptr);
-#ifdef BSL_RECORDING
-		/**
-		 * Internal use. Use the static methods scanner::file_point_cloud_stream or scanner::simple_file_point_cloud_stream.
-		 *
-		 * @param istream Input stream with Point Cloud Recording. The file format is described in the technical documentation: "Blickfeld Scanner Library : File Format".
-		 */
-		point_cloud_stream(std::istream* istream);
-#endif
-		virtual ~point_cloud_stream();
-
-#ifdef BSL_RECORDING
-		/**
-		 * Records the frames received with recv_frame to the provided ostream.
-		 * The ostream must be kept in scope until stop_recording is called.
-		 *
-		 * @param ostream Output stream for point cloud file. The file format is described in the technical documentation: "Blickfeld Scanner Library : File Format".
-		 */
-		void record_to_stream(std::ostream* ostream, int compression_level = 1);
-
-		/**
-		 * Stops the recording, which was started with record_to_stream.
-		 * Flushes the provided ostream.
-		 */
-		void stop_recording();
-
-		/**
-		 * Check if end of stream has been reached.
-		 * Only works if point_cloud_stream has been initialized with an stream input.
-		 *
-		 * @return False if further frames can be fetch with recv_frame. True if the end of stream has been reached.
-		 */
-		bool end_of_stream();
-#endif
-
-		/**
-		 * Receive frame from device or file.
-		 * Blocks until frame next frame is available.
-		 *
-		 * @return Frame Point Cloud Frame.
-		 */
-		const frame_t& recv_frame();
-
-		/** Frame subscription callback for asynchronous subscribes with subscribe. **/
-		using subscribe_callback_t = std::function<void (const frame_t&)>;
-
-		/**
-		 * Asynchronously subscribe to frames.
-		 * @param cb Callback handler for frames.
-		 */
-		void subscribe(subscribe_callback_t cb);
-
-#ifndef BSL_STANDALONE
-		/**
-		 * Return metadata of the current stream.
-		 *
-		 * @return Metadata The metadata format is described in the technical documentation: "Blickfeld Scanner Library : File Format".
-		 */
-		const protocol::file::PointCloud_Metadata get_metadata();
-#endif
-	};
+	using point_cloud_stream = point_cloud_stream<frame_t>;
 
 #ifndef BSL_STANDALONE
 	using subscribe_status_callback_t = std::function<void (const protocol::Status&)>;
@@ -189,6 +85,8 @@ protected:
 #ifdef HAVE_OPENSSL
 	asio::ssl::context* ssl_context = nullptr;
 #endif
+
+	void __hello(protocol::Response_Hello* hello);
 
 	/**
 	 * Internal use. Use static method connect.
@@ -278,7 +176,23 @@ public:
 	 */
 	void attempt_error_recovery();
 
+	/**
+	 * > Introduced in BSL v2.16 and firmware v1.17
+	 *
+	 * Retrieve unique serial number of the device.
+	 * Consider using hello() to retrieve more information.
+	 */
+	const std::string serial_number();
+
 #ifndef BSL_STANDALONE
+	/**
+	 * > Introduced in BSL v2.16 and firmware v1.17
+	 *
+	 * Retrieve product type.
+	 * Consider using hello() to retrieve more information.
+	 */
+	const protocol::config::Product product();
+
 	/**
 	 * Fetches point cloud frames from the device.
 	 * This call already boots up the device.
@@ -362,6 +276,12 @@ public:
 	 */
 	const protocol::Status get_status();
 	void subscribe(subscribe_status_callback_t cb);
+
+	/**
+	 * Send hello to device and retrieve device, software, and hardware information
+	 * @return The hello definiton can be found in the protocol section of the technical documentation.
+	 */
+	const protocol::Response::Hello hello();
 
 	/// Use this function to get the current <a href="../protobuf_protocol.html#blickfeld/config/scan_pattern.proto">ScanPattern</a>
 	const protocol::config::ScanPattern get_scan_pattern();