+++ protocol/plugin/blickfeld/error.custom.pb.h
@@ -12,7 +12,7 @@ namespace blickfeld {
 namespace protocol {
 
     static inline protocol_exception<Error::Unknown> CreateErrorUnknown(std::string description) { auto err = Error::Unknown(); err.set_description(description);  return protocol_exception<decltype(err)>(err); }
-static inline protocol_exception<Error::NotImplemented> CreateErrorNotImplemented() { auto err = Error::NotImplemented();  return protocol_exception<decltype(err)>(err); }
+static inline protocol_exception<Error::NotImplemented> CreateErrorNotImplemented(std::string reason="No detailed reason available.") { auto err = Error::NotImplemented(); err.set_reason(reason);  return protocol_exception<decltype(err)>(err); }
 static inline protocol_exception<Error::Empty> CreateErrorEmpty() { auto err = Error::Empty();  return protocol_exception<decltype(err)>(err); }
 static inline protocol_exception<Error::ServerImplementation> CreateErrorServerImplementation() { auto err = Error::ServerImplementation();  return protocol_exception<decltype(err)>(err); }
 static inline protocol_exception<Error::InvalidRequest> CreateErrorInvalidRequest(std::string validation_error) { auto err = Error::InvalidRequest(); err.set_validation_error(validation_error);  return protocol_exception<decltype(err)>(err); }
@@ -29,6 +29,7 @@ static inline protocol_exception<Error::UnknownErrorCode> CreateErrorUnknownErro
 static inline protocol_exception<Error::NotInRange> CreateErrorNotInRange(std::string parameter, float minimum, float maximum, float requested, std::string unit="") { auto err = Error::NotInRange(); err.set_parameter(parameter); err.set_minimum(minimum); err.set_maximum(maximum); err.set_requested(requested); err.set_unit(unit);  return protocol_exception<decltype(err)>(err); }
 static inline protocol_exception<Error::TimeSyncFailed> CreateErrorTimeSyncFailed(std::string ntp_daemon_log) { auto err = Error::TimeSyncFailed(); err.set_ntp_daemon_log(ntp_daemon_log);  return protocol_exception<decltype(err)>(err); }
 static inline protocol_exception<Error::NoDeviceDiscovered> CreateErrorNoDeviceDiscovered() { auto err = Error::NoDeviceDiscovered();  return protocol_exception<decltype(err)>(err); }
+static inline protocol_exception<Error::NotSupported> CreateErrorNotSupported(std::string reason="No detailed reason available.") { auto err = Error::NotSupported(); err.set_reason(reason);  return protocol_exception<decltype(err)>(err); }
 
 
     static inline void throw_typed_protocol_exception (Error err) {
@@ -51,6 +52,7 @@ case 21: throw protocol_exception<Error::UnknownErrorCode>(err.unknown_error_cod
 case 22: throw protocol_exception<Error::NotInRange>(err.not_in_range()); break; 
 case 23: throw protocol_exception<Error::TimeSyncFailed>(err.time_sync_failed()); break; 
 case 24: throw protocol_exception<Error::NoDeviceDiscovered>(err.no_device_discovered()); break; 
+case 25: throw protocol_exception<Error::NotSupported>(err.not_supported()); break; 
 
             case Error::ERROR_NOT_SET:
             default: