+++ doc/CHANGELOG.md
@@ -13,6 +13,13 @@ influence the resulting point cloud.
 
 ### Removed
 
+## [2.16.0] - 2020.12.16
+
+### Added
+* [Introduced in firmware v1.17] API: Introduce packed frame format. The format returns point clouds as flat binary structures. These can be efficiently encoded & decoded, which increases the performance on low-performance clients significantly.
+* Python: Add structured numpy interface for point clouds
+* [Introduced in firmware v1.17] Extended hello message with library, firmware, and hardware information. Added requests for serial number & product variant.
+
 ## [2.15.0] - 2020.12.02
 
 ### Added